Extract AiEditModal and BooleanStatusIcon reusable components
The four AI editor modals (ActionType, Report, Script, Tool) shared ~80%
identical code for the modal shell, chat panel, message rendering, and
footer. Extract a shared AiEditModal that owns all chat state and UI,
letting each caller provide only its unique content panel and API call.

Also extract BooleanStatusIcon for the repeated check/minus icon pattern
used in table columns across ActionTypesPage and EventSourcesPage.

Net effect: 293 lines added, 883 removed.
