+++ ui/src/pages/ActionTypesPage.tsx
@@ -20,8 +20,6 @@ import {
     Title,
 } from "@patternfly/react-core";
 import { Table, Tbody, Td, Th, Thead, Tr } from "@patternfly/react-table";
-import CheckCircleIcon from "@patternfly/react-icons/dist/esm/icons/check-circle-icon";
-import MinusCircleIcon from "@patternfly/react-icons/dist/esm/icons/minus-circle-icon";
 import PlusCircleIcon from "@patternfly/react-icons/dist/esm/icons/plus-circle-icon";
 import TrashIcon from "@patternfly/react-icons/dist/esm/icons/trash-icon";
 import {
@@ -31,6 +29,7 @@ import {
     createActionType,
     deleteActionType,
 } from "../config/api";
+import { BooleanStatusIcon } from "../components/BooleanStatusIcon";
 import { ConfirmDeleteModal } from "../components/ConfirmDeleteModal";
 
 export function ActionTypesPage() {
@@ -125,15 +124,9 @@ export function ActionTypesPage() {
                                             {at.executionMode}
                                         </Label>
                                     </Td>
-                                    <Td>{at.userTriggerable
-                                        ? <CheckCircleIcon color="var(--pf-t--global--color--status--success--default)" />
-                                        : <MinusCircleIcon color="var(--pf-t--global--icon--color--disabled)" />}</Td>
-                                    <Td>{at.managerTriggerable
-                                        ? <CheckCircleIcon color="var(--pf-t--global--color--status--success--default)" />
-                                        : <MinusCircleIcon color="var(--pf-t--global--icon--color--disabled)" />}</Td>
-                                    <Td>{at.emitsEvent
-                                        ? <CheckCircleIcon color="var(--pf-t--global--color--status--success--default)" />
-                                        : <MinusCircleIcon color="var(--pf-t--global--icon--color--disabled)" />}</Td>
+                                    <Td><BooleanStatusIcon value={at.userTriggerable} /></Td>
+                                    <Td><BooleanStatusIcon value={at.managerTriggerable} /></Td>
+                                    <Td><BooleanStatusIcon value={at.emitsEvent} /></Td>
                                     <Td>{at.executionMode === "actor" ? `${at.allowedTools?.length || 0} tools` : "—"}</Td>
                                     <Td>{at.executionMode === "actor"
                                         ? (at.promptTemplate ? "Configured" : "—")
